service collection is ready
