+++ src/PlayFramework/Rystem.PlayFramework/Chat/Models/ChatPriceSettings.cs
@@ -1,0 +1,8 @@
+﻿namespace Rystem.PlayFramework
+{
+    public sealed class ChatPriceSettings
+    {
+        public decimal InputToken { get; set; }
+        public decimal OutputToken { get; set; }
+    }
+}
