+++ src/PlayFramework/Rystem.PlayFramework/Chat/Models/Tools/ToolArrayProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Rystem.PlayFramework
+{
+    public sealed class ToolArrayProperty : ToolProperty
+    {
+        [JsonPropertyName("items")]
+        public ToolProperty? Items { get; set; }
+    }
+}
