+++ src/PlayFramework/Rystem.PlayFramework/Chat/Models/Tools/ToolProperty.cs
@@ -1,0 +1,21 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Rystem.PlayFramework
+{
+    [JsonDerivedType(typeof(ToolEnumProperty))]
+    [JsonDerivedType(typeof(ToolNumberProperty))]
+    [JsonDerivedType(typeof(ToolNonPrimitiveProperty))]
+    [JsonDerivedType(typeof(ToolArrayProperty))]
+    public class ToolProperty
+    {
+        [JsonPropertyName("type")]
+        public string Type { get; set; }
+        [JsonPropertyName("description")]
+        public string? Description { get; set; }
+        private const string DefaultTypeName = "string";
+        public ToolProperty()
+        {
+            Type = DefaultTypeName;
+        }
+    }
+}
