+++ src/PlayFramework/Rystem.PlayFramework/Interfaces/Builder/ISceneServiceBuilder.cs
@@ -1,0 +1,8 @@
+﻿namespace Rystem.PlayFramework
+{
+    public interface ISceneServiceBuilder<T>
+        where T : class
+    {
+        ISceneServiceBuilder<T> WithMethod(Func<T, Delegate> method);
+    }
+}
