+++ src/PlayFramework/Rystem.PlayFramework/Interfaces/Builder/IScenesBuilder.cs
@@ -1,4 +1,4 @@
-﻿namespace Rystem.OpenAi.Actors
+﻿namespace Rystem.PlayFramework
 {
     public interface IScenesBuilder
     {
