+++ src/PlayFramework/Rystem.PlayFramework/Interfaces/Chat/IChatClient.cs
@@ -1,0 +1,8 @@
+﻿namespace Rystem.PlayFramework
+{
+    public interface IChatClient : IChatClientBuilder, IChatClientToolBuilder
+    {
+        IAsyncEnumerable<ChatResponse> ExecuteStreamAsync(CancellationToken cancellationToken = default);
+        Task<ChatResponse> ExecuteAsync(CancellationToken cancellationToken = default);
+    }
+}
