+++ src/PlayFramework/Rystem.PlayFramework/Interfaces/Chat/IChatClientBuilder.cs
@@ -1,0 +1,11 @@
+﻿namespace Rystem.PlayFramework
+{
+    public interface IChatClientBuilder
+    {
+        IChatClient AddUserMessage(string message);
+        IChatClient AddSystemMessage(string message);
+        IChatClient AddAssistantMessage(string message);
+        IChatClient AddMaxOutputToken(int maxToken);
+        IChatClient AddPriceModel(ChatPriceSettings priceSettings);
+    }
+}
