+++ src/PlayFramework/Rystem.PlayFramework/Interfaces/Manager/ISceneManager.cs
@@ -1,4 +1,4 @@
-﻿namespace Rystem.OpenAi.Actors
+﻿namespace Rystem.PlayFramework
 {
     public interface ISceneManager
     {
