+++ src/PlayFramework/Rystem.PlayFramework/Manager/HttpBringer.cs
@@ -1,11 +1,15 @@
 ﻿using System.Text;
 
-namespace Rystem.OpenAi.Actors
+namespace Rystem.PlayFramework
 {
     internal sealed class HttpBringer
     {
         public string Method { get; set; }
         public StringBuilder Query { get; internal set; }
         public string BodyAsJson { get; internal set; }
+    }
+    internal sealed class ServiceBringer
+    {
+        public required List<object> Parameters { get; set; }
     }
 }
