+++ src/PlayFramework/Rystem.PlayFramework/Manager/SceneManagerOpenAiSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace Rystem.OpenAi.Actors
+﻿namespace Rystem.PlayFramework
 {
     public sealed class SceneManagerOpenAiSettings
     {
