+++ src/PlayFramework/Rystem.PlayFramework/Models/AiSceneResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace Rystem.OpenAi.Actors
+﻿namespace Rystem.PlayFramework
 {
     public sealed class AiSceneResponse
     {
