+++ src/PlayFramework/Rystem.PlayFramework/Rystem.PlayFramework.csproj
@@ -49,8 +49,4 @@
 		<None Update="README.md" Pack="true" PackagePath="\" />
 	</ItemGroup>
 
-	<ItemGroup>
-	  <Folder Include="Chat\" />
-	</ItemGroup>
-
 </Project>
