+++ src/PlayFramework/Rystem.PlayFramework/SwaggerFilters/ActorsOpenAiFilter.cs
@@ -5,7 +5,7 @@ using Microsoft.OpenApi.Models;
 using Rystem.PlayFramework;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace Rystem.OpenAi.Actors
+namespace Rystem.PlayFramework
 {
     public class ActorsOpenAiFilter : IOperationFilter
     {
@@ -42,8 +42,8 @@ namespace Rystem.OpenAi.Actors
                         }
                     }
                 }
-                ScenesBuilderHelper.Actions.Add(name, new());
-                ScenesBuilderHelper.Calls.Add(name, (httpBringer) =>
+                ScenesBuilderHelper.HttpActions.Add(name, new());
+                ScenesBuilderHelper.HttpCalls.Add(name, (httpBringer) =>
                 {
                     httpBringer.Method = context.ApiDescription.HttpMethod!;
                     return ValueTask.CompletedTask;
@@ -51,52 +51,24 @@ namespace Rystem.OpenAi.Actors
                 foreach (var parameter in context.ApiDescription.ParameterDescriptions)
                 {
                     var parameterName = parameter.Name ?? parameter.Type.Name;
-                    Add(parameterName, parameter.Type, jsonFunctionObject);
+                    ToolPropertyHelper.Add(parameterName, parameter.Type, jsonFunctionObject);
                     if (parameter.Source == BindingSource.Query)
                     {
-                        ScenesBuilderHelper.Actions[name][parameterName] = async (value, httpBringer) =>
+                        ScenesBuilderHelper.HttpActions[name][parameterName] = (value, httpBringer) =>
                         {
                             if (httpBringer.Query is null)
                                 httpBringer.Query = new();
                             httpBringer.Query.Append($"{parameterName}={value[parameterName]}&");
+                            return ValueTask.CompletedTask;
                         };
                     }
                     else if (parameter.Source == BindingSource.Body)
                     {
-                        ScenesBuilderHelper.Actions[name][parameterName] = (value, httpBringer) =>
+                        ScenesBuilderHelper.HttpActions[name][parameterName] = (value, httpBringer) =>
                         {
                             httpBringer.BodyAsJson = value[parameterName];
                             return ValueTask.CompletedTask;
                         };
-                    }
-
-
-                }
-                void Add(string? parameterName, Type type, ToolNonPrimitiveProperty jsonFunction)
-                {
-                    var description = type.GetCustomAttribute(typeof(DescriptionAttribute)) as DescriptionAttribute;
-                    if (type.IsPrimitive())
-                    {
-                        jsonFunction.AddPrimitive(parameterName ?? type.Name, new ToolProperty
-                        {
-                            Description = description?.Description ?? parameterName ?? type.Name,
-                            Type = type.IsNumeric() ? "number" : "string"
-                        });
-                    }
-                    else
-                    {
-                        var innerFunction = new ToolNonPrimitiveProperty()
-                        {
-                            Description = description?.Description ?? parameterName ?? type.Name,
-                        };
-                        jsonFunction.AddObject(parameterName ?? type.Name, innerFunction);
-                        foreach (var innerParameter in type.GetProperties(BindingFlags.Instance | BindingFlags.Public))
-                        {
-                            if (innerParameter.GetCustomAttribute<System.Text.Json.Serialization.JsonIgnoreAttribute>() is null)
-                            {
-                                Add(innerParameter.Name, innerParameter.PropertyType, innerFunction);
-                            }
-                        }
                     }
                 }
             }
